add functions to adding targets to the bacground additively
This new feature adds possibility to render the trajectory with specified images for objects on the given background. ggplot can't add additively (e.g. blend the image) to the background, therefore we need to add the targets into the background and show the full image instead of ggplot objects.

Limitations:
- There are no sanity chekcs in the moments (Issue: #1)
- Only square backgrounds are allowed (Issue: #2)
- Current version does work only with same image for all object (Issue: #3)
-